==> 02_Scripts/metabolome/04_mouse_serum_tissue_sphingolipids.rmd <==
---
title: "Functional role of SMPD1 in PDAC"
subtitle: "__Significantly altered metabolites (Figure 3)__"
author: "_umahajan_"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_notebook:
    theme: united
    number_sections: yes
    toc: yes
    toc_depth: 4
    toc_float: false
---

```{r setup, include=FALSE}
chooseCRANmirror(graphics=TRUE, ind=1)
knitr::opts_chunk$set(tidy.opts=list(width.cutoff=85),tidy=TRUE, echo=TRUE, warning=FALSE, message=FALSE)
```

# load packages and datasets

```{r packages}
rm(list = ls())
##----------------------------------------------------------------
##                        load packages                        --
##----------------------------------------------------------------
scriptLibraries <-  c(
  "here",
  "tidyverse",
  "readxl",
  "metapacR",
  "RColorBrewer",
  "ggpubr",
  "rstatix"
)

#devtools::install_github("umahajanatlmu/metapacR", ref="master", auth_token = "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")

##---------------------------------------------------------------
##                      load functions                         --
##---------------------------------------------------------------
devtools::source_url("https://raw.githubusercontent.com/umahajanatlmu/useful_commands/main/auxillary/basicFunctions.R")

# load packages 
##---------------------------------------------------------------
installScriptLibs(scriptLibraries)

## results directory
ifelse(!dir.exists(file.path(paste0(here()), "results_metabolome_mm")),
       dir.create(file.path(paste0(here()), "results_metabolome_mm")), FALSE)

save_plot <- function(x, filename, width = 7, height = 7) {
    stopifnot(!missing(x))
    stopifnot(!missing(filename))
    svg(filename, width = width, height = height)
    grid::grid.newpage()
    grid::grid.draw(x)
    dev.off()
}
```

# untargeted analysis (mouse)

## load data

```{r}
## metabolite name
serum_met <- readxl::read_excel("../../Data/Metabolome/mouse/MTXH_30_METAPAC_09_Mouse_Serum_ANOVA_GROUP.xlsx",
                                    skip=1) %>%
  mutate(MET_CHEM_NO=as.character(MET_CHEM_NO)) %>%
  dplyr::select(MET_CHEM_NO, ONTOLOGY1_NAME, ONTOLOGY2_NAME, METABOLITE_NAME) 

## pull cer metabolites
serum_cer_met <- readxl::read_excel("../../Data/Metabolome/mouse/MTXH_30_METAPAC_09_Mouse_Serum_ANOVA_GROUP.xlsx",
                                    skip=1) %>%
  filter(ONTOLOGY2_NAME =="Ceramides") %>%
  mutate(MET_CHEM_NO=as.character(MET_CHEM_NO)) %>%
  pull(MET_CHEM_NO) 

## pull sm metabolites
serum_sm_met <- readxl::read_excel("../../Data/Metabolome/mouse/MTXH_30_METAPAC_09_Mouse_Serum_ANOVA_GROUP.xlsx",
                                    skip=1) %>%
  filter(ONTOLOGY2_NAME =="Sphingomyelins") %>%
  mutate(MET_CHEM_NO=as.character(MET_CHEM_NO)) %>%
  pull(MET_CHEM_NO)

## load dataset
serum_dat <- read.csv("../../Data/Metabolome/mouse/MTXH_30_METAPAC_09_Mouse_Serum_DATA.txt", 
                      sep = "\t", colClasses = "character", header = TRUE) %>%
  filter(grepl("^C\\d|^PDAC",EXTERNAL_ID)) %>%
  column_to_rownames("EXTERNAL_ID") %>%
  dplyr::select(starts_with("X")) 

colnames(serum_dat) <- gsub("^X", "", colnames(serum_dat))

## subset cer
serum_dat_cer <- serum_dat[, grepl("^1.30037E.12", colnames(serum_dat)) ]
colnames(serum_dat_cer) <- serum_cer_met

## subset sm
serum_dat_sm <- serum_dat[, grepl("^1.10034E.12", colnames(serum_dat)) ]
colnames(serum_dat_sm) <- serum_sm_met

## knit serum_dat
serum_dat <- serum_dat[, !grepl("^1.30037E.12|^1.10034E.12", colnames(serum_dat)) ]


chains <- c("C14", "C16", "C18", "C20", "C22", "C24", "C26")

serum_dat_cumm <- bind_cols(serum_dat, serum_dat_cer, serum_dat_sm) %>%
  rownames_to_column("EXTERNAL_ID") %>%
  metapacR::ImputeTransformScale(.,
                                 Impute=TRUE,
                                 Transform=TRUE,
                                 Scaling=TRUE,
                                 ScaleType = "Auto") %>%
  mutate(across(where(is.numeric), ~replace_na(., median(., na.rm=TRUE)))) %>%
  pivot_longer(cols = -EXTERNAL_ID, 
               names_to = "MET_CHEM_NO",
               values_to = "value") %>%
  left_join(serum_met, by="MET_CHEM_NO" ) %>%    
  filter(ONTOLOGY1_NAME == "Complex lipids, fatty acids and related") %>%
  filter(ONTOLOGY2_NAME == "Ceramides" | ONTOLOGY2_NAME == "Sphingomyelins") %>%
  filter(grepl("^CER_|^SM_",METABOLITE_NAME)) %>%
  filter(grepl("d18:1,",METABOLITE_NAME))  %>% ## only d18:1
  filter(grepl(paste(chains, collapse = "|"),METABOLITE_NAME))  %>%
  dplyr::select(ONTOLOGY2_NAME, EXTERNAL_ID, value) %>%
  group_by(ONTOLOGY2_NAME, EXTERNAL_ID) %>%
  summarise(median_value = median(as.numeric(value), na.rm = TRUE)) %>%
  ungroup() %>%
  mutate_if(is.numeric, funs(. / .[ONTOLOGY2_NAME == "Sphingomyelins"])) %>%
  filter(ONTOLOGY2_NAME != "Sphingomyelins") %>%
  mutate(MATRIX="Serum") %>%
  dplyr::select(-ONTOLOGY2_NAME)
```

```{r}
## metabolite name
tissue_met <- readxl::read_excel("../../Data/Metabolome/mouse/MTXH_30_METAPAC_09_Mouse_Pancreas_ANOVA_GROUP.xlsx",
                                    skip=1) %>%
  mutate(MET_CHEM_NO=as.character(MET_CHEM_NO)) %>%
  dplyr::select(MET_CHEM_NO, ONTOLOGY1_NAME, ONTOLOGY2_NAME, METABOLITE_NAME) 

## pull cer metabolites
tissue_cer_met <- readxl::read_excel("../../Data/Metabolome/mouse/MTXH_30_METAPAC_09_Mouse_Pancreas_ANOVA_GROUP.xlsx",
                                    skip=1) %>%
  filter(ONTOLOGY2_NAME =="Ceramides") %>%
  mutate(MET_CHEM_NO=as.character(MET_CHEM_NO)) %>%
  pull(MET_CHEM_NO) 

tissue_cer_met <- tissue_cer_met[stringr::str_length(tissue_cer_met) == 13]


## pull sm metabolites
tissue_sm_met <- readxl::read_excel("../../Data/Metabolome/mouse/MTXH_30_METAPAC_09_Mouse_Pancreas_ANOVA_GROUP.xlsx",
                                    skip=1) %>%
  filter(ONTOLOGY2_NAME =="Sphingomyelins") %>%
  mutate(MET_CHEM_NO=as.character(MET_CHEM_NO)) %>%
  pull(MET_CHEM_NO)

tissue_sm_met <- tissue_sm_met[stringr::str_length(tissue_sm_met) == 13]

## load dataset
tissue_dat <- read.csv("../../Data/Metabolome/mouse/MTXH_30_METAPAC_09_Mouse_Pancreas_DATA.tsv", 
                      sep = "\t", colClasses = "character", header = TRUE) %>%
  filter(grepl("^C\\d|^PDAC",EXTERNAL_ID)) %>%
  column_to_rownames("EXTERNAL_ID") %>%
  dplyr::select(starts_with("X")) 

colnames(tissue_dat) <- gsub("^X", "", colnames(tissue_dat))

## subset cer
tissue_dat_cer <- tissue_dat[, grepl("^1.30037E.12", colnames(tissue_dat)) ]
colnames(tissue_dat_cer) <- tissue_cer_met

## subset sm
tissue_dat_sm <- tissue_dat[, grepl("^1.10034E.12", colnames(tissue_dat)) ]
colnames(tissue_dat_sm) <- tissue_sm_met

## knit serum_dat
tissue_dat <- tissue_dat[, !grepl("^1.30037E.12|^1.10034E.12", colnames(tissue_dat)) ]


chains <- c("C14", "C16", "C18", "C20", "C22", "C24", "C26")

tissue_dat_cumm <- bind_cols(tissue_dat, tissue_dat_cer, tissue_dat_sm) %>%
  rownames_to_column("EXTERNAL_ID") %>%
  metapacR::ImputeTransformScale(.,
                                 Impute=TRUE,
                                 Transform=TRUE,
                                 Scaling=TRUE,
                                 ScaleType = "Auto") %>%
  mutate(across(where(is.numeric), ~replace_na(., median(., na.rm=TRUE)))) %>%
  pivot_longer(cols = -EXTERNAL_ID, 
               names_to = "MET_CHEM_NO",
               values_to = "value") %>%
  left_join(tissue_met, by="MET_CHEM_NO" ) %>%    
  filter(ONTOLOGY1_NAME == "Complex lipids, fatty acids and related") %>%
  filter(ONTOLOGY2_NAME == "Ceramides" | ONTOLOGY2_NAME == "Sphingomyelins") %>%
  filter(grepl("^CER_|^SM_",METABOLITE_NAME)) %>%
  filter(grepl("d18:1,",METABOLITE_NAME))  %>% ## only d18:1
  filter(grepl(paste(chains, collapse = "|"),METABOLITE_NAME))  %>%
  dplyr::select(ONTOLOGY2_NAME, EXTERNAL_ID, value) %>%
  group_by(ONTOLOGY2_NAME, EXTERNAL_ID) %>%
  summarise(median_value = median(as.numeric(value), na.rm = TRUE)) %>%
  ungroup() %>%
  mutate_if(is.numeric, funs(. / .[ONTOLOGY2_NAME == "Sphingomyelins"])) %>%
  filter(ONTOLOGY2_NAME != "Sphingomyelins") %>%
  mutate(MATRIX="Tissue") %>%
  dplyr::select(-ONTOLOGY2_NAME)
```

## CER/SM ratio

```{r}
dat_cumm <- bind_rows(serum_dat_cumm, tissue_dat_cumm) %>%
  mutate(EXTERNAL_ID = gsub("[^CPDAC]", "", EXTERNAL_ID))

stat.test <- dat_cumm %>% 
  group_by(MATRIX) %>%
  wilcox_test(median_value ~ EXTERNAL_ID) %>%
  add_xy_position(
          x = "EXTERNAL_ID",
          step.increase = 0.15
        )

p <- ggplot(dat_cumm,
            aes(x=EXTERNAL_ID,
                y=median_value,
                fill=EXTERNAL_ID)
            ) +
  geom_boxplot(show.legend = FALSE) +
  geom_jitter(
          shape = 21,
          size = 2,
          width = 0.2,
          color = "black",
          alpha = 0.3,
          show.legend = FALSE
        ) +
  stat_pvalue_manual(
          #y = 0.5,
          stat.test,
          label = "p",
          tip.length = 0.01,
          color = "gray50",
          size = 4,
          bracket.size = 0.6,
          inherit.aes = FALSE
        ) +
  xlab("") +
  ylab("CER/SM ratio") +
  theme_bw() +
  theme(
    axis.line = element_line(size = 0.5),
    axis.text = element_text(
      size = 11,
      face = "bold",
      colour = "black"
    ),
    axis.title = element_text(size = 12, face = "bold")
  ) +
  scale_fill_manual(values=c("#377EB8","#E41A1C")) +
  # expand_limits(y=3.5) +
  # scale_y_continuous(trans = scales::log10_trans(),
  #   breaks = scales::trans_breaks("log10", function(x) 10^x),
  #   labels = scales::trans_format("log10", scales::math_format(10^.x))) +
  facet_grid(~MATRIX)

print(p)

# sjPlot::save_plot(filename = paste0("./results_metapac_vd2/CER_SM_ratio.svg"),
#                   fig = p,
#                   width = 8,
#                   height = 7,
#                   dpi = 300)

ggsave("./results_metabolome_mm/CER_SM_ratio.pdf",
         plot = p, 
         width = 9,
         height = 8,
         dpi = 300,
         units = "cm")

```

## CER/SM ratio metabologram

```{r}
serum_dat_chain <- bind_cols(serum_dat, serum_dat_cer, serum_dat_sm) %>%
  rownames_to_column("EXTERNAL_ID") %>%
  metapacR::ImputeTransformScale(.,
                                 Impute=TRUE,
                                 Transform=TRUE,
                                 Scaling=TRUE,
                                 ScaleType = "Auto") %>%
  pivot_longer(cols = -EXTERNAL_ID, 
               names_to = "MET_CHEM_NO",
               values_to = "value") %>%
  left_join(serum_met, by="MET_CHEM_NO" ) %>%    
  filter(ONTOLOGY1_NAME == "Complex lipids, fatty acids and related") %>%
  filter(ONTOLOGY2_NAME == "Ceramides" | ONTOLOGY2_NAME == "Sphingomyelins") %>%
  filter(grepl("^CER_|^SM_",METABOLITE_NAME)) %>%
  filter(grepl("d18:1,",METABOLITE_NAME))  %>% ## only d18:1
  filter(grepl(paste(chains, collapse = "|"),METABOLITE_NAME))  %>%
  mutate(METABOLITE_NAME2 = gsub(".*,", "", METABOLITE_NAME)) %>%
  mutate(METABOLITE_NAME2 = gsub(")", "", METABOLITE_NAME2, fixed=TRUE)) %>%
  group_by(ONTOLOGY2_NAME, EXTERNAL_ID, METABOLITE_NAME2) %>%
  summarise(median_value = median(as.numeric(value), na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(MATRIX="Serum")


tissue_dat_chain <- bind_cols(tissue_dat, tissue_dat_cer, tissue_dat_sm) %>%
  rownames_to_column("EXTERNAL_ID") %>%
  metapacR::ImputeTransformScale(.,
                                 Impute=TRUE,
                                 Transform=TRUE,
                                 Scaling=TRUE,
                                 ScaleType = "Auto") %>%
  pivot_longer(cols = -EXTERNAL_ID, 
               names_to = "MET_CHEM_NO",
               values_to = "value") %>%
  left_join(serum_met, by="MET_CHEM_NO" ) %>%    
  filter(ONTOLOGY1_NAME == "Complex lipids, fatty acids and related") %>%
  filter(ONTOLOGY2_NAME == "Ceramides" | ONTOLOGY2_NAME == "Sphingomyelins") %>%
  filter(grepl("^CER_|^SM_",METABOLITE_NAME)) %>%
  filter(grepl("d18:1,",METABOLITE_NAME))  %>% ## only d18:1
  filter(grepl(paste(chains, collapse = "|"),METABOLITE_NAME))  %>%
  mutate(METABOLITE_NAME2 = gsub(".*,", "", METABOLITE_NAME)) %>%
  mutate(METABOLITE_NAME2 = gsub(")", "", METABOLITE_NAME2, fixed=TRUE)) %>%
  group_by(ONTOLOGY2_NAME, EXTERNAL_ID, METABOLITE_NAME2) %>%
  summarise(median_value = median(as.numeric(value), na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(MATRIX="Tissue")
```

```{r}
dat_chain <- bind_rows(serum_dat_chain, tissue_dat_chain) %>% 
  mutate(EXTERNAL_ID = gsub("[^CPDAC]", "", EXTERNAL_ID)) %>%
  group_by(MATRIX, EXTERNAL_ID) %>%
  mutate_if(is.numeric, funs(. / .[ONTOLOGY2_NAME == "Sphingomyelins"])) %>%
  filter(ONTOLOGY2_NAME != "Sphingomyelins") %>% 
  ungroup() %>%
  dplyr::select(-ONTOLOGY2_NAME)

stat.test <- dat_chain %>% 
  mutate(Group = interaction(EXTERNAL_ID, METABOLITE_NAME2)) %>%
  group_by(MATRIX) %>%
  wilcox_test(median_value ~ Group) %>%
  filter(grepl("^C.", group1) & grepl("^PDAC.", group2)) %>%
  mutate(C_group = gsub(".*\\.", "", group1),
         PDAC_group = gsub(".*\\.", "", group2)) %>%
  filter(C_group == PDAC_group) %>%
  ungroup()

dat_chain_summary <- dat_chain %>%
  group_by(METABOLITE_NAME2, MATRIX) %>%
  mutate_if(is.numeric, funs(. / .[EXTERNAL_ID == "C"])) %>%
  filter(EXTERNAL_ID != "C") %>% 
  summarise(ratio=median(median_value, na.rm=TRUE)) %>%
  ungroup() %>%
  inner_join(stat.test, by = c("MATRIX", "METABOLITE_NAME2" = "PDAC_group")) %>%
  dplyr::select(METABOLITE_NAME2, MATRIX, ratio, p)

unique_met <- unique(dat_chain_summary$METABOLITE_NAME2)

dat_chain_summary_c <- dat_chain_summary %>%
  arrange(MATRIX,METABOLITE_NAME2) %>%
  mutate(ceramide_x = rep(seq(1, length(unique_met), by=1),2),
         sphingomyelin_x = rep(seq(length(unique_met) + 3 , (2*length(unique_met)) +2, by=1),2),
         curvature = c(runif(length(unique_met), min = 0.3, max = 0.8),runif(length(unique_met), min = -0.8, max = -0.3)) )

# Create a color palette with "RdBu" colors
color_palette <- colorRampPalette(brewer.pal(length(unique_met), "RdBu"))(length(unique_met))

# Filter the data for 'Tissue'
tissue_data <- dat_chain_summary_c[dat_chain_summary_c$MATRIX == "Tissue", ]
# Sort the colors based on your specified order
tissue_data$color <- color_palette[order(tissue_data$ratio)]

# Filter the data for 'Serum'
serum_data <- dat_chain_summary_c[dat_chain_summary_c$MATRIX == "Serum", ]
# Sort the colors based on your specified order
serum_data$color <- color_palette[order(serum_data$ratio)]
```

```{r}
# Define a function to create a geom_curve layer
create_curve_layer <- function(data_row) {
  geom_curve(data = data_row,
             aes(x = ceramide_x,
                 xend = sphingomyelin_x,
                 y = ifelse(MATRIX == "Tissue", 2, 1.5),
                 yend = ifelse(MATRIX == "Tissue", 2, 1.5),
                 size = data_row[["ratio"]]),
             color=data_row[["color"]],
             alpha = 0.7,
             curvature = data_row[["curvature"]])
}

# Filter the data for 'Tissue' and create a list of 'geom_curve' layers
curve_layers_tissues <- lapply(1:nrow(tissue_data), function(i) {
  create_curve_layer(tissue_data[i, ])
})

# Filter the data for 'Serum' and create a list of 'geom_curve' layers
curve_layers_serum <- lapply(1:nrow(serum_data), function(i) {
  create_curve_layer(serum_data[i, ])
})


p <- ggplot() + 
  theme_classic() +
  labs(x = "", y = "") +
  curve_layers_tissues +
  curve_layers_serum +
  geom_rect(aes(xmin = 0, 
                xmax = length(unique_met) + 1, 
                ymin = 1.45, 
                ymax = 2.05),
            fill = "#377EB8", 
            color = "#000000") +
  geom_text(aes(x = (0.5 + length(unique_met) / 2), 
                y = 2, 
                label = "Ceramide", 
                face = "bold"), 
            color = "white") +
  geom_rect(aes(xmin = length(unique_met) + 2, 
                xmax = 2 * length(unique_met) + 3, 
                ymin = 1.45,
                ymax = 2.05),
            fill = "#E41A1C", 
            color = "#000000") +
  geom_text(aes(x = (1 + length(unique_met) * 1.65), 
                y = 2, 
                label = "Sphingomyelin", 
                face = "bold"), 
            color = "white") +
  scale_x_continuous(breaks = seq(1, 2 * length(unique_met) + 2, by = 1), 
                     labels = c(unique_met, "", "", unique_met)) +
  scale_y_continuous(limits = c(0, 3.2)) +
  theme(axis.text.y = element_blank(),
        axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5, face = "bold"),
        axis.line = element_blank(),
        axis.ticks = element_blank(),
        legend.position = "none")

print(p)

ggsave("./results_metabolome_mm/CER_SM_metabologram.pdf",
         plot = p, 
         width = 15,
         height = 11,
         dpi = 300,
         units = "cm")

```


# computing environment

```{r}
sessionInfo()
```
